ingestion/batch/ingest_fx/pyproject.toml
[project]
name = "financial-transactions-analytics-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "freecurrencyapi>=0.1.0",
    "google-cloud-storage>=2.10.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]
